spn: utils: cover kmeans single cluster corner case
This patch fixes an issue with k-means returning a single cluster when
k>1. This occurred when all instances were sufficiently similar to each
other, meaning one cluster would have the minimal distance to every
point. This is fixed by using a variation of muesli/kmeans.Partition,
renatogeh/kmeans.PartitionPoints.

The original function Partition returns a slice containing each cluster
and their designated observations. This is an inefficient way of
clustering in our use-case, as a posterior assigning of the ordered
original observations would take O(n^2). PartitionPoints takes care of
this issue by simply returning the slice of predictions in order.

Due to the severity of this bug, this patch includes renatogeh/kmeans
until the relevant pull request is accepted into muesli/kmeans.

Signed-off-by: Renato Lui Geh <renatogeh@gmail.com>